split demote_alloca and underlying ptr analysis
